refactor: move optional type
